Styling changes and switched ModifyGameDetails fields to drop-down menus
