--- src/admin-client/GameInsert.tsx
@@ -1,198 +1,0 @@
-import React, { useState } from 'react';
-import { useMutation, useQuery } from '@apollo/react-hooks';
-import { makeStyles } from '@material-ui/core/styles';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import DialogContent from '@material-ui/core/DialogContent';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-import Select from '@material-ui/core/Select';
-import InputLabel from '@material-ui/core/InputLabel';
-import MenuItem from '@material-ui/core/MenuItem';
-import FormControl from '@material-ui/core/FormControl';
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from '@material-ui/pickers';
-import DateFnsUtils from '@date-io/date-fns';
-import { timeSelections, currentLeagueTimes } from '../utils/time';
-import { GAME_SCHEDULE_INSERT } from '../graphql/mutations/game.mutations';
-import { TEAM_DETAILS_QUERY } from '../graphql/queries/team.queries';
-
-function GameInsert(props: {refetchGames?: () => void, closeMenu?: () => void}) {
-
-    const classes = useStyles();
-
-    const { data } = useQuery(TEAM_DETAILS_QUERY);
-
-    const [insertGame] = useMutation(GAME_SCHEDULE_INSERT);
-
-    const [awayTeam, setAwayTeam] = useState<string>();
-    const [homeTeam, setHomeTeam] = useState<string>();
-    const [leagueYear, setLeagueYear] = useState<number>(currentLeagueTimes.currentLeagueYear);
-    const [gameTime, setGameTime] = useState<string>(currentLeagueTimes.usualGameTime);
-    const [gameWeek, setGameWeek] = useState<number>(currentLeagueTimes.currentLeagueWeek);
-    const [gameDate, setGameDate] = useState<Date>(new Date());
-
-    const teamOptions = data?.team_details.map((team) => team.name);
-
-    const handleGameInsert = () => {
-        insertGame(
-            {
-                variables: {
-                    team_1_name: awayTeam, 
-                    team_2_name: homeTeam, 
-                    league_year: leagueYear, 
-                    time: gameTime + ':00', 
-                    week: gameWeek , 
-                    date: gameDate
-                }
-            }
-        )
-
-        if (props.refetchGames) {
-            props.refetchGames()
-        };
-
-        if (props.closeMenu) {
-            props.closeMenu()
-        };
-    };
-
-    return (
-        <div>
-            <DialogTitle>
-                Create Game
-            </DialogTitle>
-            <DialogContent dividers>
-                <DialogContent className={classes.inputRow}>
-                    <DialogContent>
-                        <FormControl className={classes.teamSelectors}>
-                            <InputLabel htmlFor="away-team-id">Away Team</InputLabel>
-                            <Select
-                                value={awayTeam}
-                                labelId="away-team-id"
-                                onChange={(fieldValue) => setAwayTeam(fieldValue.target.value as string)}
-                            >
-                                {data && (
-                                    teamOptions.map((name) => {
-                                        return <MenuItem value={name}>{name}</MenuItem>
-                                    })
-                                )}
-                            </Select>
-                        </FormControl>
-                    </DialogContent>
-                    
-                    <DialogContent>
-                        <FormControl className={classes.teamSelectors}>
-                            <InputLabel htmlFor="home-team-id">Home Team</InputLabel>
-                            <Select
-                                value={homeTeam}
-                                labelId="home-team-id"
-                                onChange={(fieldValue) => setHomeTeam(fieldValue.target.value as string)}
-                            >
-                                {data && (
-                                    teamOptions.map((name) => {
-                                        return <MenuItem value={name}>{name}</MenuItem>
-                                    })
-                                )}
-                            </Select>
-                        </FormControl>
-                    </DialogContent>
-                </DialogContent>
-
-                <DialogContent className={classes.inputRow}>
-                    <DialogContent>
-                        <Typography>
-                            Date
-                        </Typography>
-                        <MuiPickersUtilsProvider utils={DateFnsUtils}>
-                            <div> 
-                                <KeyboardDatePicker
-                                    disableToolbar
-                                    margin="normal"
-                                    format="yyyy-MM-dd"
-                                    value={gameDate}
-                                    onChange={(date) => setGameDate(date as Date)}
-                                />
-                            </div>
-                        </MuiPickersUtilsProvider>
-                    </DialogContent>
-
-                    <DialogContent>
-                        <Typography>
-                            Time (CST)
-                        </Typography>
-                        <TextField
-                            type="time"
-                            onChange={(fieldValue) => setGameTime(fieldValue.target.value)}
-                            inputProps={{
-                                step: 300,
-                            }}
-                        />
-                    </DialogContent>
-                </DialogContent>
-                
-                <DialogContent className={classes.inputRow}>
-                    <DialogContent>
-                        <FormControl className={classes.teamSelectors}>
-                            <InputLabel htmlFor="week-id">Week</InputLabel>
-                            <Select
-                                value={gameWeek}
-                                labelId="week-id"
-                                onChange={(fieldValue) => setGameWeek(fieldValue.target.value as number)}
-                            >
-                                {
-                                    timeSelections.leagueWeeks.map((week) => {
-                                        return <MenuItem value={week}>{week}</MenuItem>
-                                    })
-                                }
-                            </Select>
-                        </FormControl>
-                    </DialogContent>
-
-                    <DialogContent>
-                        <FormControl className={classes.teamSelectors}>
-                            <InputLabel htmlFor="year-id">Year</InputLabel>
-                            <Select
-                                value={leagueYear}
-                                labelId="year-id"
-                                onChange={(fieldValue) => setLeagueYear(fieldValue.target.value as number)}
-                            >
-                                {
-                                    timeSelections.leagueYears.map((year) => {
-                                        return <MenuItem value={year}>{year}</MenuItem>
-                                    })
-                                }
-                            </Select>
-                        </FormControl>
-                    </DialogContent>
-                </DialogContent>
-
-                <DialogContent className={classes.buttonRow}>
-                    <Button variant="outlined" onClick={handleGameInsert} className={classes.addButton}>
-                        Add Game
-                    </Button>
-                </DialogContent>
-            </DialogContent>
-        </div>
-    );
-}
-
-export default GameInsert;
-
-const useStyles = makeStyles({
-    inputRow: {
-        display: 'flex',
-    },
-    teamSelectors: {
-        display: 'flex',
-    },
-    addButton: {
-        textTransform: 'none',
-    },
-    buttonRow: {
-        display: 'flex', 
-        justifyContent: 'flex-end',
-    },
-});
